Reincorporación de la agenda

+++ src/AppBundle/Controller/EventoController.php
@@ -3,47 +3,29 @@
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\Evento;
+use http\Env\Response;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * Evento controller.
  *
- * @Route("/evento")
+ * @Route("/agenda")
  */
 class EventoController extends Controller
 {
     /**
-     * @Route("/", name="marina-agenda")
+     * @Route("/", name="evento_index")
      */
     public function displayMarinaAgenda(Request $request)
     {
         $em = $this->getDoctrine()->getManager();
-        $marinaEventos = $em->getRepository('AppBundle:Evento')->findAll();
-
+        $eventos = $em->getRepository('AppBundle:Evento')->eventosVisibles($this->getUser());
         return $this->render('evento/marina-agenda.twig', [
-            'marinaEventos' => $marinaEventos,
-            'marinaagenda' => 1
+            'marinaEventos' => $eventos,
         ]);
-    }
-
-    /**
-     * Lists all evento entities.
-     *
-     * @Route()
-     * @Method("GET")
-     */
-    public function indexAction()
-    {
-        $em = $this->getDoctrine()->getManager();
-
-        $eventos = $em->getRepository('AppBundle:Evento')->findAll();
-
-        return $this->render('evento/index.html.twig', array(
-            'eventos' => $eventos,
-            'marinaagenda' => 1
-        ));
     }
 
     /**
@@ -55,22 +37,22 @@ class EventoController extends Controller
     public function newAction(Request $request)
     {
         $evento = new Evento();
+        $this->denyAccessUnlessGranted('AGENDA_CREATE',$evento);
         $form = $this->createForm('AppBundle\Form\EventoType', $evento);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
+            $evento->setUsuario($this->getUser());
             $em->persist($evento);
             $em->flush();
-
-            return $this->redirectToRoute('evento_show', array('id' => $evento->getId()));
+            return $this->redirectToRoute('evento_show', [
+                'id' => $evento->getId()
+            ]);
         }
-
-        return $this->render('evento/new.html.twig', array(
+        return $this->render('evento/new.html.twig', [
             'evento' => $evento,
-            'form' => $form->createView(),
-            'marinaagenda' => 1
-        ));
+            'form' => $form->createView()
+        ]);
     }
 
     /**
@@ -81,13 +63,17 @@ class EventoController extends Controller
      */
     public function showAction(Evento $evento)
     {
-        $deleteForm = $this->createDeleteForm($evento);
-
-        return $this->render('evento/show.html.twig', array(
+        $editable = false;
+        if($this->getUser() === $evento->getUsuario() || $this->getUser()->isAdmin()){
+            $editable = true;
+        }
+        if(!$evento->getIsPublico() && ($this->getUser() !== $evento->getUsuario())){
+            throw new NotFoundHttpException();
+        }
+        return $this->render('evento/show.html.twig', [
             'evento' => $evento,
-            'delete_form' => $deleteForm->createView(),
-            'marinaagenda' => 1
-        ));
+            'editable' => $editable
+        ]);
     }
 
     /**
@@ -98,22 +84,19 @@ class EventoController extends Controller
      */
     public function editAction(Request $request, Evento $evento)
     {
+        $this->denyAccessUnlessGranted('AGENDA_EDIT',$evento);
         $deleteForm = $this->createDeleteForm($evento);
         $editForm = $this->createForm('AppBundle\Form\EventoType', $evento);
         $editForm->handleRequest($request);
-
         if ($editForm->isSubmitted() && $editForm->isValid()) {
             $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('evento_show', array('id' => $evento->getId()));
+            return $this->redirectToRoute('evento_show', ['id' => $evento->getId()]);
         }
-
-        return $this->render('evento/edit.html.twig', array(
+        return $this->render('evento/edit.html.twig', [
             'evento' => $evento,
             'edit_form' => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
-            'marinaagenda' => 1
-        ));
+        ]);
     }
 
     /**
@@ -124,15 +107,14 @@ class EventoController extends Controller
      */
     public function deleteAction(Request $request, Evento $evento)
     {
+        $this->denyAccessUnlessGranted('AGENDA_DELETE',$evento);
         $form = $this->createDeleteForm($evento);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($evento);
             $em->flush();
         }
-
         return $this->redirectToRoute('evento_index');
     }
